+++ .gitignore
@@ -14,3 +14,5 @@ test-results.xml
 test_config.json
 .idea
 testUserData
+
+.gitattributes
