+++ package.json
@@ -53,6 +53,7 @@
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.4",
     "chai": "^4.2.0",
+    "copy-webpack-plugin": "^6.2.1",
     "cross-env": "^5.2.0",
     "css-loader": "^1.0.1",
     "devtron": "^1.4.0",
