+++ src/main/tray/tray.js
@@ -1,0 +1,67 @@
+// Copyright (c) 2016-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import path from 'path';
+import {app, nativeImage, nativeTheme} from 'electron';
+
+const assetsDir = path.resolve(app.getAppPath(), 'renderer/assets');
+
+export function getTrayImages(trayIconTheme) {
+  switch (process.platform) {
+  case 'win32':
+    return {
+      normal: nativeImage.createFromPath(path.resolve(assetsDir, 'windows/tray.ico')),
+      unread: nativeImage.createFromPath(path.resolve(assetsDir, 'windows/tray_unread.ico')),
+      mention: nativeImage.createFromPath(path.resolve(assetsDir, 'windows/tray_mention.ico')),
+    };
+  case 'darwin':
+  {
+    const icons = {
+      light: {
+        normal: nativeImage.createFromPath(path.resolve(assetsDir, 'osx/MenuIcon.png')),
+        unread: nativeImage.createFromPath(path.resolve(assetsDir, 'osx/MenuIconUnread.png')),
+        mention: nativeImage.createFromPath(path.resolve(assetsDir, 'osx/MenuIconMention.png')),
+      },
+      clicked: {
+        normal: nativeImage.createFromPath(path.resolve(assetsDir, 'osx/ClickedMenuIcon.png')),
+        unread: nativeImage.createFromPath(path.resolve(assetsDir, 'osx/ClickedMenuIconUnread.png')),
+        mention: nativeImage.createFromPath(path.resolve(assetsDir, 'osx/ClickedMenuIconMention.png')),
+      },
+    };
+    switchMenuIconImages(icons, nativeTheme.shouldUseDarkColors);
+    return icons;
+  }
+  case 'linux':
+  {
+    const theme = trayIconTheme;
+    try {
+      return {
+        normal: nativeImage.createFromPath(path.resolve(assetsDir, 'linux', theme, 'MenuIconTemplate.png')),
+        unread: nativeImage.createFromPath(path.resolve(assetsDir, 'linux', theme, 'MenuIconUnreadTemplate.png')),
+        mention: nativeImage.createFromPath(path.resolve(assetsDir, 'linux', theme, 'MenuIconMentionTemplate.png')),
+      };
+    } catch (e) {
+      //Fallback for invalid theme setting
+      return {
+        normal: nativeImage.createFromPath(path.resolve(assetsDir, 'linux', 'light', 'MenuIconTemplate.png')),
+        unread: nativeImage.createFromPath(path.resolve(assetsDir, 'linux', 'light', 'MenuIconUnreadTemplate.png')),
+        mention: nativeImage.createFromPath(path.resolve(assetsDir, 'linux', 'light', 'MenuIconMentionTemplate.png')),
+      };
+    }
+  }
+  default:
+    return {};
+  }
+}
+
+export function switchMenuIconImages(icons, isDarkMode) {
+  if (isDarkMode) {
+    icons.normal = icons.clicked.normal;
+    icons.unread = icons.clicked.unread;
+    icons.mention = icons.clicked.mention;
+  } else {
+    icons.normal = icons.light.normal;
+    icons.unread = icons.light.unread;
+    icons.mention = icons.light.mention;
+  }
+}
